# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Dispatch API
  version: 3.0.0
  description: |
    Open-source bulk email campaign platform with tracking, multi-provider support, and batch processing.

    ## Authentication
    Most endpoints require authentication via a `session_token` cookie obtained from the login endpoint.
    Alternatively, an API key can be passed via the `X-API-Key` header.

    ## CSRF Protection
    All mutation endpoints (POST, PUT, DELETE) require a valid `X-CSRF-Token` header.
    The CSRF token is issued as a cookie on every response.

    ## Common Response Format
    All endpoints return a standard envelope:
    ```json
    {
      "success": true,
      "data": { ... },
      "message": "Human-readable message",
      "meta": { "pagination": { ... } }
    }
    ```
  contact:
    name: Dispatch
    url: https://github.com/iamjpsingh/bulk-email-sender
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
      description: Session token cookie set on login/register.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for programmatic access.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF token required for POST/PUT/DELETE requests. Value is read from the `csrf_token` cookie.

  # -------------------------------------------------------------------------
  # Reusable Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          description: Response payload (shape varies per endpoint).
        message:
          type: string
        meta:
          type: object
          properties:
            pagination:
              $ref: '#/components/schemas/Pagination'
      required: [success]
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Something went wrong
        error:
          type: string
      required: [success, message]
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5
        hasMore:
          type: boolean
          example: true

    # -- Auth --
...
    SMTPConfigUpdate:
      type: object
      properties:
        name:
          type: string
        host:
          type: string
        port:
          type: integer
        secure:
          type: boolean
        user:
          type: string
        pass:
          type: string
          format: password
        fromEmail:
          type: string
          format: email
        fromName:
          type: string
        isDefault:
          type: boolean
...
    JobSummary:
      type: object
      properties:
        id:
          type: string
        campaign_id:
          type: string
          nullable: true
        type:
          type: string
          enum: [direct, batch, scheduled, automation]
        status:
          type: string
          enum: [pending, running, paused, completed, failed, cancelled]
        priority:
          type: integer
        subject:
          type: string
        from_email:
          type: string
        config_name:
          type: string
        total_count:
          type: integer
        sent_count:
          type: integer
        failed_count:
          type: integer
        last_processed_index:
          type: integer
        batch_size:
          type: integer
        progress:
          type: integer
          description: Percentage 0-100.
        scheduled_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        last_error:
          type: string
          nullable: true
...
    SuppressionEntry:
      type: object
      properties:
        email:
          type: string
          format: email
        reason:
          type: string
        source:
          type: string
        created_at:
          type: string
          format: date-time

    # -- Reports --
...
  parameters:
...
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
    OffsetParam:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
...
security:
  - cookieAuth: []
  - apiKeyAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /config/smtp/{configId}:
    put:
      tags: [Config]
      summary: Update SMTP configuration
      operationId: updateSmtpConfig
      parameters:
        - name: configId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SMTPConfigUpdate'
      responses:
        '200':
          description: Configuration updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Configuration not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [Config]
      summary: Delete SMTP configuration
      operationId: deleteSmtpConfig
      parameters:
        - name: configId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Configuration deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Configuration not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /queue/jobs:
    get:
      tags: [Queue]
      summary: List jobs
      operationId: listJobs
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, paused, completed, failed, cancelled]
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Array of job summaries.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/JobSummary'
  /queue/jobs/{id}:
    get:
      tags: [Queue]
      summary: Get a specific job
      operationId: getJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job details.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/JobSummary'
        '404':
          description: Job not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [Queue]
      summary: Cancel a job
      operationId: cancelJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job cancelled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Job cannot be cancelled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /queue/jobs/{id}/pause:
    post:
      tags: [Queue]
      summary: Pause a running job
      operationId: pauseJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job paused.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Job is not running.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /queue/jobs/{id}/resume:
    post:
      tags: [Queue]
      summary: Resume a paused job
      operationId: resumeJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job resumed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Job is not paused.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /queue/dead-letters:
    get:
      tags: [Queue]
      summary: List dead letter entries
      operationId: listDeadLetters
      parameters:
        - name: job_id
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Dead letter entries.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
  /queue/suppression:
    get:
      tags: [Queue]
      summary: Get suppression list
      operationId: getSuppressionList
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Suppression list entries.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/SuppressionEntry'

    post:
      tags: [Queue]
      summary: Add email to suppression list
      operationId: suppressEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                reason:
                  type: string
              required: [email, reason]
      responses:
        '200':
          description: Email suppressed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Email and reason are required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /queue/suppression/{email}:
    delete:
      tags: [Queue]
      summary: Remove email from suppression list
      operationId: unsuppressEmail
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: Email unsuppressed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Email not found in suppression list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # =========================================================================
  # Reports
  # =========================================================================
...